requirements.txt
arcade==2.4.1
attrs==20.1.0
cffi==1.14.2
numpy==1.19.1
Pillow==7.2.0
pycparser==2.20
pyglet==1.5.7
pymunk==5.6.0
pytiled-parser==0.9.4a3
PyYAML==5.3.1
Shapely~=1.7.1